refactor: InputText를 컴포넌트로 사용하여 style을 분기 처리 하도록 수정
