--- src/components/InputDescription/InputDescription.module.css
@@ -1,5 +1,0 @@
-.input-description {
-  font-size: 9.5px;
-  font-weight: 500;
-  color: var(--color-description);
-}
