--- src/components/InputSubtitle/InputSubtitle.module.css
@@ -1,5 +1,0 @@
-.input-subtitle {
-  font-size: 12px;
-  font-weight: 500;
-  padding: 18px 0 8px 0;
-}
