--- src/components/InputTitle/InputTitle.module.css
@@ -1,5 +1,0 @@
-.input-title {
-  font-size: 18px;
-  font-weight: 700;
-  padding: 38px 0 5px 0;
-}
